[refactor] 북마크 로직 삭제 및 리뷰 로직 위치 변경
