+++ src/main/java/com/example/OMEB/domain/user/application/service/UserService.java
@@ -1,23 +1,16 @@
 package com.example.OMEB.domain.user.application.service;
 
-import com.example.OMEB.domain.book.persistence.entity.BookMark;
-import com.example.OMEB.domain.book.persistence.repository.BookMarkRepository;
-import com.example.OMEB.domain.review.persistence.entity.Review;
-import com.example.OMEB.domain.review.persistence.repository.ReviewRepository;
 import com.example.OMEB.domain.user.application.IncreaseExpType;
 import com.example.OMEB.domain.user.persistence.entity.ExpLog;
 import com.example.OMEB.domain.user.persistence.entity.User;
 import com.example.OMEB.domain.user.persistence.repository.ExpLogRepository;
 import com.example.OMEB.domain.user.persistence.repository.UserRepository;
-import com.example.OMEB.domain.user.presentation.dto.response.UserBookMarkResponse;
 import com.example.OMEB.domain.user.presentation.dto.response.UserExpLogResponse;
 import com.example.OMEB.domain.user.presentation.dto.response.UserInfoResponse;
-import com.example.OMEB.domain.user.presentation.dto.response.UserReviewResponse;
 import com.example.OMEB.domain.user.presentation.dto.response.rank.UserRankInfoResponse;
 import com.example.OMEB.domain.user.presentation.dto.response.rank.UserRankPageResponse;
 import com.example.OMEB.global.base.exception.ErrorCode;
 import com.example.OMEB.global.base.exception.ServiceException;
-import com.nimbusds.openid.connect.sdk.claims.UserInfo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -31,9 +24,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class UserService {
     private final UserRepository userRepository;
-    private final ReviewRepository reviewRepository;
     private final ExpLogRepository expLogRepository;
-    private final BookMarkRepository bookMarkRepository;
 
     public UserInfoResponse getUserInfo(Long userId){
         User user = userRepository.findById(userId)
@@ -41,25 +32,13 @@ public class UserService {
         return UserInfoResponse.entityToResponse(user);
     }
 
-    public List<UserReviewResponse> getUserReviews(Long userId){
-        List<Review> reviewList = reviewRepository.findByUser_id(userId);
 
-        return reviewList.stream()
-                .map(UserReviewResponse::entityToResponse).toList();
-    }
 
     public List<UserExpLogResponse> getUserExpLogs(Long userId){
         List<ExpLog> expLogList = expLogRepository.findByUser_id(userId);
 
         return expLogList.stream()
                 .map(UserExpLogResponse::entityToResponse).toList();
-    }
-
-    public List<UserBookMarkResponse> getUserBookMarks(Long userId){
-        List<BookMark> bookMarkList = bookMarkRepository.findByUser_id(userId);
-
-        return bookMarkList.stream()
-                .map(UserBookMarkResponse::entityToResponse).toList();
     }
 
     @Transactional
@@ -75,8 +54,7 @@ public class UserService {
         return UserRankPageResponse.pageToResponse(userRankListResponsePage);
     }
 
-    @Transactional
-    public ExpLog increaseExp(User user, IncreaseExpType type){
+    public void increaseExp(User user, IncreaseExpType type){
         int increasedExp = user.getExp() + type.getExp();
         int level = user.getLevel();
         int levelExpSize = 100 * (int)Math.pow(2, level-1);
@@ -86,10 +64,11 @@ public class UserService {
             user.updateLevel(level);
         }
         user.updateExp(increasedExp);
-        return ExpLog.builder()
+
+        expLogRepository.save(ExpLog.builder()
                 .exp(type.getExp())
                 .content(type.getContent())
                 .user(user)
-                .build();
+                .build());
     }
 }
